Move the rest of manageContent actions
