+++ kolibri/plugins/management/assets/src/state/actions.js
@@ -1,13 +1,10 @@
 const coreApp = require('kolibri');
-const logging = require('kolibri.lib.logging');
 const getters = require('kolibri.coreVue.vuex.getters');
 
 const ClassroomResource = coreApp.resources.ClassroomResource;
 const MembershipResource = coreApp.resources.MembershipResource;
 const FacilityUserResource = coreApp.resources.FacilityUserResource;
-const TaskResource = coreApp.resources.TaskResource;
 const RoleResource = coreApp.resources.RoleResource;
-const ChannelResource = coreApp.resources.ChannelResource;
 
 const coreActions = require('kolibri.coreVue.vuex.actions');
 const ConditionalPromise = require('kolibri.lib.conditionalPromise');
@@ -15,7 +12,6 @@ const constants = require('../constants');
 const UserKinds = require('kolibri.coreVue.vuex.constants').UserKinds;
 
 const PageNames = constants.PageNames;
-const ContentWizardPages = constants.ContentWizardPages;
 const samePageCheckGenerator = require('kolibri.coreVue.vuex.actions').samePageCheckGenerator;
 // because these modules use ES6 module syntax, need to access exports.default in CommonJS context
 const addCoachRoleAction = require('./addCoachRoleAction').default;
@@ -92,18 +88,6 @@ function _userState(apiUserData) {
     full_name: apiUserData.full_name,
     kind: calcUserKind(apiUserData.roles),
   };
-}
-
-
-function _taskState(data) {
-  const state = {
-    id: data.id,
-    type: data.type,
-    status: data.status,
-    metadata: data.metadata,
-    percentage: data.percentage,
-  };
-  return state;
 }
 
 /**
@@ -488,171 +472,6 @@ function showUserPage(store) {
   );
 }
 
-
-// ================================
-// CONTENT IMPORT/EXPORT ACTIONS
-
-function updateWizardLocalDriveList(store) {
-  const localDrivesPromise = TaskResource.localDrives();
-  store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', true);
-  localDrivesPromise.then((response) => {
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', false);
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_DRIVES', response.entity);
-  })
-  .catch((error) => {
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', false);
-    coreActions.handleApiError(store, error);
-  });
-}
-
-function updateWizardPage(store, pageName, meta = {}) {
-  return store.dispatch('SET_CONTENT_PAGE_WIZARD_STATE', {
-    busy: false,
-    driveList: null,
-    drivesLoading: false,
-    error: null,
-    page: pageName,
-    meta,
-    shown: true,
-  });
-}
-
-function startImportWizard(store) {
-  return updateWizardPage(store, ContentWizardPages.CHOOSE_IMPORT_SOURCE);
-}
-
-function startExportWizard(store) {
-  updateWizardLocalDriveList(store);
-  return updateWizardPage(store, ContentWizardPages.EXPORT);
-}
-
-function showImportNetworkWizard(store) {
-  return updateWizardPage(store, ContentWizardPages.IMPORT_NETWORK);
-}
-
-function showLocalImportPreview(store, driveData) {
-  return updateWizardPage(store, ContentWizardPages.IMPORT_PREVIEW, {
-    sourceId: driveData.driveId,
-    sourceName: driveData.driveName,
-    sourceType: 'local',
-    channels: driveData.channels,
-    error: null,
-  });
-}
-
-function showNetworkImportPreview(store, channelId) {
-  const channelMetadataRequest = ChannelResource.getModel(channelId).fetch();
-  const payload = {
-    sourceId: channelId,
-    sourceName: '',
-    sourceType: 'network',
-    channels: [], // just to make types consistent
-    error: null,
-  };
-  return channelMetadataRequest._promise
-  .then((channel) => {
-    updateWizardPage(
-      store,
-      ContentWizardPages.IMPORT_PREVIEW,
-      Object.assign(payload, { sourceName: channel.name })
-    );
-  })
-  .catch((err) => {
-    updateWizardPage(
-      store,
-      ContentWizardPages.IMPORT_PREVIEW,
-      Object.assign(payload, { error: err.entity })
-    );
-  });
-}
-
-function showImportLocalWizard(store) {
-  updateWizardLocalDriveList(store);
-  return updateWizardPage(store, ContentWizardPages.IMPORT_LOCAL);
-}
-
-function cancelImportExportWizard(store) {
-  return updateWizardPage(store, ContentWizardPages.NONE);
-}
-
-// called from a timer to continually update UI
-function pollTasksAndChannels(store) {
-  const samePageCheck = samePageCheckGenerator(store);
-  TaskResource.getCollection().fetch({}, true).only(
-    // don't handle response if we've switched pages or if we're in the middle of another operation
-    () => samePageCheck() && !store.state.pageState.wizardState.busy,
-    (taskList) => {
-      // Perform channel poll AFTER task poll to ensure UI is always in a consistent state.
-      // I.e. channel list always reflects the current state of ongoing task(s).
-      coreActions.setChannelInfo(store).only(
-        samePageCheckGenerator(store),
-        () => {
-          store.dispatch('SET_CONTENT_PAGE_TASKS', taskList.map(_taskState));
-          // Close the wizard if there's an outstanding task.
-          // (this can be removed when we support more than one
-          // concurrent task.)
-          if (taskList.length && store.state.pageState.wizardState.shown) {
-            cancelImportExportWizard(store);
-          }
-        }
-      );
-    },
-    error => { logging.error(`poll error: ${error}`); }
-  );
-}
-
-function clearTask(store, taskId) {
-  const clearTaskPromise = TaskResource.clearTask(taskId);
-  clearTaskPromise.then(() => {
-    store.dispatch('SET_CONTENT_PAGE_TASKS', []);
-  })
-  .catch(error => { coreActions.handleApiError(store, error); });
-}
-
-function triggerLocalContentImportTask(store, driveId) {
-  store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', true);
-  const localImportPromise = TaskResource.localImportContent(driveId);
-  localImportPromise.then((response) => {
-    store.dispatch('SET_CONTENT_PAGE_TASKS', [_taskState(response.entity)]);
-    cancelImportExportWizard(store);
-  })
-  .catch((error) => {
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_ERROR', error.status.text);
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', false);
-  });
-}
-
-function triggerLocalContentExportTask(store, driveId) {
-  store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', true);
-  const localExportPromise = TaskResource.localExportContent(driveId);
-  localExportPromise.then((response) => {
-    store.dispatch('SET_CONTENT_PAGE_TASKS', [_taskState(response.entity)]);
-    cancelImportExportWizard(store);
-  })
-  .catch((error) => {
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_ERROR', error.status.text);
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', false);
-  });
-}
-
-function triggerRemoteContentImportTask(store, channelId) {
-  store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', true);
-  const remoteImportPromise = TaskResource.remoteImportContent(channelId);
-  remoteImportPromise.then((response) => {
-    store.dispatch('SET_CONTENT_PAGE_TASKS', [_taskState(response.entity)]);
-    cancelImportExportWizard(store);
-  })
-  .catch((error) => {
-    if (error.status.code === 404) {
-      store.dispatch('SET_CONTENT_PAGE_WIZARD_ERROR', 'That ID was not found on our server.');
-    } else {
-      store.dispatch('SET_CONTENT_PAGE_WIZARD_ERROR', error.status.text);
-    }
-    store.dispatch('SET_CONTENT_PAGE_WIZARD_BUSY', false);
-  });
-}
-
-
 // ================================
 // OTHER ACTIONS
 
@@ -698,20 +517,6 @@ const actions = {
   showUserPage,
   addCoachRole: addCoachRoleAction,
   removeCoachRole: removeCoachRoleAction,
-
-  pollTasksAndChannels,
-  clearTask,
-  startImportWizard,
-  startExportWizard,
-  showImportNetworkWizard,
-  showImportLocalWizard,
-  showLocalImportPreview,
-  showNetworkImportPreview,
-  cancelImportExportWizard,
-  triggerLocalContentExportTask,
-  triggerLocalContentImportTask,
-  triggerRemoteContentImportTask,
-  updateWizardLocalDriveList,
 
   showDataPage,
   showScratchpad,
